Fix: Resolution now properly shows in the GUI after setting the value

+++ src/MODEXngine.renderlib.opengl/OpenGLRenderer.cs
@@ -49,7 +49,7 @@ namespace MODEXngine.renderlib.opengl
             Height = a.Height,
             Width = a.Width,
             Bpp = a.BitsPerPixel,
-            RefreshRate = a.RefreshRate
+            RefreshRate = (int) a.RefreshRate
         }).ToList();
     }
 }
